requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
chat2plot==0.0.11
click==8.1.7
colorama==0.4.6
commentjson==0.9.0
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.0.3
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.9
langchain-community==0.2.7
langchain-core==0.2.21
langchain-text-splitters==0.2.2
langsmith==0.1.90
lark-parser==0.7.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.35.14
openpyxl==3.1.5
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.22.0
protobuf==5.27.2
psutil==6.0.0
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
streamlit==1.36.0
streamlit-chat==0.1.1
tabulate==0.9.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vegafusion==1.6.9
vegafusion-python-embed==1.6.9
vl-convert-python==1.5.0
watchdog==4.0.1
Werkzeug==3.0.3
yarl==1.9.4
